Revert log change for applied_fork_recovery_block
In the commits "Switch HTTP handlers logging to new utils", "Switch
ar_node_worker to the new logging utils" and "Switch fork recovery
logs to the new utils" some ar:report_console/1 and ar:report/1 calls
got changed to the new ar:console/1, ar:warn/1 and ar:err/1 calls that
I don't longer aggree with.

Some ar:err/1 call were just plain wrong since it's an expected
behaviour where there is code to deal with it, so it's better just log
an INFO or WARNING message.

We probably would like to replace the ar:report_console/1 calls with
ar:console/1 in the future, but first improve the output of
ar:console/1 for when the argument is a proplist-ish.