Fix bug when miners reject their own blocks
The bug occurs very rarely when a new transaction arrives into the mempool (passing the network propagation delay)
shortly after the miner mines a block. The weave size is by mistake computed from the ready for mining transactions
instead of mined transactions what leads to a validation failure and rejection of the block.