Do not "fork recover" to height+1 blocks
The motivation is to disincentivize submission of globally rejected content. In an unwanted scenario when the whole network rejects a particular chunk of illicit data, adding it to the network should have a cost. One extra block is the cost we choose. A tolerable downside of it is in a scenario when a half (very roughly) of the network rejects a chunk the network is going to be more forky than it could.