Switch from SHA-384 to RandomX PoW hashing at fork 1.7
RandomX is a PoW algorithm that is optimized for general-purpose
CPUs. RandomX uses random code execution together with several
memory-hard techniques to maximize the efficiency of general-purpose
CPUs and therefore make GPU accelerated miners and ASIC based miners
less likely to be advantageous.

RandomX is a C++ library with a C interface. We're using a NIF module
(ar_mine_randomx) for gluing it together with the Erlang code. To
reduce the time the NIF calls are blocking the scheduler, the work is
performed in one or more threads and the result is then sent in a
message to the calling Erlang process.

The RandomX version used is a commit between the 1.0.2 and 1.0.3
release where one of the minor security issues discovered by an audit
has been addressed (more AES rounds). On top of these, we have tweaked
the parameters according to the auditor's recommendations. The dataset
was decreased from default 2 GiB to 1 GiB to reduce the memory
requirements for miners.

Initializing a RandomX dataset is heavy work which takes roughly 7
minutes (but can be parallelized in multiple threads). The
ar_randomx_state module provides a server which makes sure to
initialize the datasets in time before they are needed and then caches
them until they are not needed anymore. At most 2 datasets are
allocated at the same time.

When running the tests (i.e. when DEBUG is set) it's too slow to
generate the datasets. Instead, we then run in "light-mode" which
generates only a RandomX cache instead of a dataset, which takes only
around 1 second to generate. The down-side is that hashing performance
drops by an order of magnitude.

Co-authored-by: Lev Berman <ldmberman@protonmail.com>