Implement perpetual TX pricing & linear difficulty
The update takes effect after the hard fork 1.8 (featuring multiple transactions per wallet per block).

The transaction price is computed based on the estimated perpetual storage costs, conservatively assuming
exponential decay of the storage costs. The estimation is done in USD and then converted into AR based on
the initial exchange rate, inflation, and the change in the network difficulty.

Estimating the exchange rate based on difficulty benefits from the latter changing smoothly, therefore the
change comes with the linear difficulty model. The difficulty is now a number the target hash must be greate than.
`2 ** 256 / (2 ** 256 - Diff)` indicates the number of hashes one has to try on average before finding a block.
Previously, it was a power of 2 of the same amount. The difficulty is now adjusted proportionally to the ratio
between the actual time and the target time for ten blocks.

The new difficulty and cumulative difficulty are stored and transferred as strings.

The cumulative difficulty is incremented by the average number of hashes required to find a solution
under the difficulty of the given block. Therefore, the new cumulative difficulty suggests the total
number of hashes the network has tried since the fork 1.8 minus the number accumulated before the fork.

Co-authored-by: Lev Berman <ldmberman@protonmail.com>