Stop banning peers sending invalid blocks
ar_node_worker might ban the peer which a propagated block came from
after the block got propagated to the node's peers. This means the
banning will continue through the network, banning all nodes' best
peers. We're still banning the peers on edge if the light PoW check is
invalid.