Introduce a GraphQL interface for TX index
This commit introduces a new interface to query data in the weave using GraphQL.

The main issues addressed by the new interface are:

- Use an existing and well defined query language to let apps fetch data from the weave
- Reduce the number of roundtrips required to get an exploitable dataset to use in an app on the permaweb
- Allow querying through user-defined relationships between transactions

GraphQL requests are accepted by the POST /arql endpoint. The ArQL language is
still supported - the requests are dispatched to the corresponding backend
according to the received query.

The new queries are resolved using the same backing database as the ArQL queries,
meaning that the feature onlt presents negligible performance implications compared
to the previous implementation.

GraphQL queries are validated against the schema defined in priv/schema.graphql.