Implement a TX propagation priority queue
Network bandwidth is a limited resource therefore miners
need a way to prioritise certain transactions over the other.
The commit introduces a priority queue where the transactions
with the higher utility are propagated first. The utility is
computed as reward / size - a proxy for the cost of a unit of
traffic. Scaling the utility this way creates a market for the
limited bandwidth resource.

The commit also changes the order in which transactions
are included into blocks - higher utility transactions are
included first. In this case, the utility function proxies
the cost per unit of storage.

Co-authored-by: Lev Berman <ldmberman@protonmail.com>