Tune RandomX
Add command line for enabling/disabling JIT, hardware AES, and large pages.

Hardware AES is supported by most modern processes. It is on by default.
Large pages need to be configured in the OS, the flag is switched off by default.
JIT is off by default on MacOS, the compilation and linking needs tweaking
to enable it - currently, it segfaults when called from the NIF shared library.