Change the disk space limit feature to be more efficient
The new behaviour is to apply the given used space limit to the disk partition with the Arweave folder, not to the folder itself. In case somebody wants to strictly conrol how much of the weave is written independently from any external writes, they need to dedicate a separate partition to the Arweave folder. The previous use-case of the per-folder control is arguably not useful and hardly popular so it is worth getting rid of the significant overhead of periodic folder traversal in Erlang.