Store one block index per node, every 50 blocks
Storing block index takes time. It is only needed in case of emergency so it does not make sense to do it too often and we do not need to store a new file every time the node is restarted. For convenience, change the startup option to "start_from_block_index", which does not accept any arguments.