Drop the "init new weave" feature
It requires extra effort to reset the fork heights and update the inflation rewards issuance, to make the
new weaves created this way function without the issues solved by the hard forks in the original weave. At the
same time, we plan to remove the outdated fork branches from the consensus code. Therefore, we choose to drop
the feature for now. There are no fundamental obstacles to reintroducing the feature in the future.