Drop committed transactions from priority queue
Normally, transactions are included in blocks only after they leave the priority queue. However, a TX may
be sent to multiple nodes and if the load distribution is not uniform, the node under a bigger pressure may
see the transaction included in a block while it still stays in its queue.