Make wallet state management & hashing scalable
The commit changes the way wallets are inserted, updated, hashed, stored, and shared with peers.

The commit includes a hard fork (2.2) required for switching to the new recursive hashing
method.

Wallets are now stored in a tree, closely resembling Patricia Merkle trees.
Each node in the tree stores the hash of its subtree, what allows to efficiently
recompute the hash of the whole collection - it scales with the number of transactions in
the block, not with the total number of wallets in the weave. Such a tree has an advantage over
the binary tree because there is a hard cap on the number of operations which may be
required for recomputing the root hash after a single insert - 32 (key length) * 2 ^ 8.
A binary tree, on the other hand, requires periodic rebalancing, otherwise the complexity
of operations may approach O(number of wallets).

Wallet trees of the recent blocks (including the blocks from forks) are kept in a DAG,
where the entire tree is only stored once, and the other trees are constructed from it
by applying the diffs from the path between the two nodes in the graph.

The change allows to store only one wallet tree at any time reducing the space
requirements to O(number of wallets in the weave) + O(number of wallets changed recently)
from the previous O(number of wallets in the weave * number of recent blocks). Moreover,
the efficiency of reconstructing a wallet tree of a previous block (e.g., when applying a fork)
becomes O(number of wallets changed since the base block of the fork), a huge improvement over
O(number of wallets in the weave), previously required when the entire tree was read from disk.

The wallet trees are still backed up on disk every block (or every several blocks,
if a fork is applied), to keep serving the entire wallet lists to the nodes running
the previous versions efficiently. However, the updated nodes running this change will attempt
to download wallet trees in chunks (and consequently serve chunks via the new
GET /wallet_list/<id>[/<cursor>] endpoint). After the majority of nodes migrates (in this case,
after the fork 2.2), the support for the old way of serving wallet lists can be dropped,
and the nodes may switch to only storing at most one wallet tree. Finally, to complete
the transition to the scalable wallet tree management, we will stop ever loading the entire tree
into RAM (currently, the ar_wallets gen_server keeps the DAG with the wallet tree in its state;
the external interface of the server is not expected to change).

The commit also includes a refactoring of the tests which spawn extra ar_node processes.
ar_node is a singleton by design so simulating multiple nodes by multiplying it does not
match the production setup of several communicating nodes. This commit further enlarges
the process graph by introducing a server managing wallet trees (ar_wallets) so it felt like
the right time to get rid of the hacky test setups. These tests now depend on the second
VM launched in the distributed mode. The internal gossip network is updated to support
sending messages to other Erlang nodes.