Set up Erlang application and supervision graph
The reason such a basic thing comes so late is it was never a priority until now
when we have two real cases for graceful and possibly long shutdown, which
can only be arranged if the application and the supervisors are set up right.

One case is closing RocksDB. The other case is persisting the mempool.

Moreover, the new setup ensures the whole VM shuts down if the core ar_node process dies at most
once or one of the supervisors gives up restarting processes (with more liberal restart policies).

Finally, this commit fixes a cyclic dependency between ar_node and ar_join, which caused startup
(and one of the tests) to take 30 seconds (?LOCAL_NET_TIMEOUT) longer.