Increase the benchmark's block time
A bigger block time allows to measure with bigger randomx_bulk_hashing_iterations, since otherwise
when the benchmark mines a block and computes the hashrate, it lacks the information about the
hashes tried by the mining threads that are still running as the number of iterations is big but
the block time is short. It matters especially on big machines with many hashing threads.