Fetch new blocks' missing transactions earlier
Before this change, the node would wait for another block to arrive before attempting
to download the missing transactions from the first block. This was an attempt to disincentivize
mining transactions into blocks without giving them reasonable time to spread across the network.

However, the disincentivization is there even without the change, because it takes time for
verifiers to download transactions, what increases the chance of a fork away from the block
in question. Therefore, the extra delay merely increases the forkiness in the network.