Move chunk data to RocksDB
The primary concern is the inefficiency and the limitations of the common file systems.
For instance, ext4 has a relatively low 4 billion max files limit and has to be tuned to maintain large directories.
The work is also a part of preparation for the new consensus mechanism where the client would have to perform
random access as fast as possible. Using a blob storage saves up a significant amount of time by reducing
the number of system calls required to read a certain data volume.

Chunk data is moved to a dedicated RocksDB instance. The keys are prefixed by the insertion time
to minimize the amount of compaction needed to maintain the database - as long as RocksDB does not change
the order of insertion drastically, the created SST files have non-overlapping key ranges making search
efficient from the start.

On top of that, the commit configures prefix bloom filters for chunks_index and switches to same-prefix lookups
to leverage them.

Bloom filters (but not prefix bloom filters) are added to data_root_index, tx_index, and disk_pool_chunks_index
because their keys are sparse.

The new DB instance for chunk data is configured with 256 MiB memtables and SST files.
The increase allows the database to keep more file descriptors open without hitting the configured system and user limits.