Re-introduce dynamic USD to AR rate adjustment
Before the hard fork 2.4, the protocol used to estimate the USD to AR exchange rate
based on how much the network difficulty has changed since the time where AR had
a certain price in USD, hard-coded in the protocol. There was a high uncertainty about
how much the difficulty would drop at the fork 2.4 where the new mining algorithm was
introduced. Obviously such a drop is not expected to correlate with the token price.

After giving the network some time to stabilize, we are re-introducing the dynamic
difficulty-based adjustment, also in a better shape than it used to be. Now the rate
to be used in the new block is determined by the previous block, what frees the miners
from the necessity to recompute the prices and the Merkle root of the accounts with
every timestamp refresh when mining retarget blocks. Additionally, the adjusted rate
does not take effect immediately that allows the /price endpoint to give an estimation
that won't expire and won't be unnecessarily big at the same time.

Two new block fields are introduced from the 2.5 fork block on: usd_to_ar_rate and
scheduled_usd_to_ar_rate. The values of both are tuples {Dividend, Divisor}, where
dividends and divisors are big integers. A tuple is serialized into a list of two
number strings in JSON.