Add a "testnet" compilation target
Reduce the block time to 1 second, the initial difficulty - to 6,
the minimal difficulty to 2, remove the base transaction propagation delay.

Fix the USD to AR adjustment function that would compute a very high rate for the early blocks
because the inflation begins at height 1, not 0.

Add POST /mine and GET /tx/ready_for_mining endpoints.