Allow for separate monitoring of data disk space
The motivation is to allow miners to mount data and headers to different disks - it makes sense
to use a faster disk for the dataset and a cheaper disk to store headers. Specifically,
the rocksdb and chunk_storage folders' disks are monitored separately from the root (data_dir)
folder disk.