Introduce a "disk cache"
The disk cache is a place where we store and rotate recent blockchain headers. The motivation is
to keep the node running under limited disk space. When the disk space is running low, the
synchronization of historical data and headers is stopped, but a number of recent blocks and
transactions are always available for serving to peers and quick joining of the network.

Previously, a rotation was performed by ar_header_sync and ar_data_sync but was not efficient due
to the limitations of RocksDB (it cannot deallocate disk space after removing the keys quickly).

Co-authored-by: Lev Berman <ldmberman@protonmail.com>