Gossip blocks via block announcements
Block announcements are small requests that help to reduce traffic.
Additionally, they provide an interface that lets the recipient specify
the missing transactions and chunks, what allows to deliver missing
transactions faster on one hand, and save traffic on not sending the
chunks on the other. The latter is only implemented in the interface -
in practice, chunks are still going to be sent for now.