Manually trigger the garbage collector periodically.
Several of the processes involved in mining "touch" chunks. The
way Erlang's memory managment works each time a process touches
a chunk it gets a reference to the full 100MiB recall range.

Also the default garbage collection frequency is based on the
amount of activity a given process does. So for some processes
the default GC may happen infrequently. This can lead to
the 100MiB recall ranges staying in memory long enough to
trigger memory issues.

By manually GC'ing periodically we ensure memory is released
regularly.