Remove most instances of process_flag(trap_exit, true)
Only keep the ones in processes that write something to disk.
I believe the use of trap_exit is part of the reason for the slow
node shutdown. I think we only need in situatinos where there's
an operation in proecss that needs to be terminated gracefully
(e.g. closing an open file)