Mining performance optimizations (#596)
With this PR we've confirmed that a single-node miner can mine the full weave (56 partitions as of today) at about 95% efficiency. We ran the tests on a 128c EPYC with SMT disabled, but Arweave only ran at 17-25% CPU utilization - this implies a much smaller CPU might also be able to mine the full weave.

When running on a large core CPU it's important to limit some of the threading. This can be done by using the `hashing_threads`, `max_vdf_validation_thread_count`, `max_vdf_last_step_validation_thread_count`, and `packing_rate` flags or by passing a flag directly to the Erlang VM 

An example using the application flags:
```
./bin/start hashing_threads 24 max_vdf_validation_thread_count 4 max_vdf_last_step_validation_thread_count 4 packing_rate 100 <other options>
```

An example using the VM flag:
```
./bin/start +S 24:24 -- <other options>
```

Note: the above restrictions may nor may not be needed - and the values used may change. You'll want to run tests on your miner to find the best options to use. The above examples were used to mostly restrict the 128-core CPU to uses 24-cores for hashing. without this restriction the thread contention from 128 threads hurt performance.

Main changes in this PR:
- Expose the Erlang VM flags from the command line. All options before a `--` separator are passed to the Erlang VM, all arguments after it are passed to Arweave. If there is no `--` then all arguments are passed to Arweave.
- Aggregate updates to `ar_mining_stats` rather than incrementing the counters with each operation. At 56 partitions the constant incrementing created an ets table bottleneck which hurt performance. Now hash and chunk updates are aggregated so that far fewer ETS operations are needed.
- Bug fix in the caching of `ar_chunk_storage` file handles. Prior to this fix the chunk storage files were opened and closed on every read because the key used in the cache differed from the key used in the lookup.
- Fix `data-doctor bench` which had broken in an earlier release
- Improve the node shutdown time by removing a call to `process_flag(trap_exit, true),`
- Slight tweak to the Ideal read/hash rate reported in the mining performance report - previously the tip partition was included in the ideal rate even though it is not minable.